refactor(vercel-ai, webhook): Rename toAiModelMessage to toModelMessage and move to vercel-ai
- Renamed  to  in .
- Moved the  function from  to .
- Updated  exports to include .
- Updated  documentation for Vercel AI SDK to use the new function name and import path.
- Created  for the new  function.
- Removed  related tests from .
