+++ packages/aikit-respond-io/package.json
@@ -14,6 +14,10 @@
     "./client": {
       "import": "./dist/client/index.js",
       "types": "./dist/client/index.d.ts"
+    },
+    "./vercel-ai": {
+      "import": "./dist/vercel-ai/index.js",
+      "types": "./dist/vercel-ai/index.d.ts"
     }
   },
   "scripts": {
