+++ packages/aikit-respond-io/tests/utils.test.ts
@@ -2,8 +2,7 @@
 import { describe, expect, test } from "bun:test";
 
 import { InvalidSignatureError } from "../src/webhook/errors";
-import { MessageContent } from "../src/webhook/types";
-import { toAiModelMessage, verifySignature } from "../src/webhook/utils";
+import { verifySignature } from "../src/webhook/utils";
 
 describe("utils", () => {
   describe("verifySignature", () => {
@@ -38,77 +37,6 @@ describe("utils", () => {
       const shortSignature = Buffer.from("short").toString("base64");
       expect(() => verifySignature(body, shortSignature, signingKey)).toThrow(
         InvalidSignatureError,
-      );
-    });
-  });
-
-  describe("toAiModelMessage", () => {
-    test("should convert text message for user", () => {
-      const content: MessageContent = { type: "text", text: "Hello" };
-      const result = toAiModelMessage(content, "user");
-
-      expect(result).toEqual({
-        role: "user",
-        content: [{ type: "text", text: "Hello" }],
-      });
-    });
-
-    test("should convert image attachment for user", () => {
-      const content: MessageContent = {
-        type: "attachment",
-        attachment: { type: "image", url: "https://example.com/image.jpg" },
-      };
-
-      const result = toAiModelMessage(content, "user");
-
-      expect(result).toEqual({
-        role: "user",
-        content: [
-          {
-            type: "image",
-            image: new URL("https://example.com/image.jpg"),
-          },
-        ],
-      });
-    });
-
-    test("should throw error for unsupported attachment type", () => {
-      const content: MessageContent = {
-        type: "attachment",
-        attachment: { type: "video", url: "https://example.com/video.mp4" },
-      } as any;
-
-      expect(() => toAiModelMessage(content, "user")).toThrow(
-        "Unsupported attachment type: video",
-      );
-    });
-
-    test("should throw error for unsupported message type", () => {
-      const content = { type: "location" } as any;
-
-      expect(() => toAiModelMessage(content, "user")).toThrow(
-        "Unsupported message type: location",
-      );
-    });
-
-    test("should convert text message for assistant", () => {
-      const content: MessageContent = { type: "text", text: "Response" };
-      const result = toAiModelMessage(content, "assistant");
-
-      expect(result).toEqual({
-        role: "assistant",
-        content: "Response",
-      });
-    });
-
-    test("should throw error for non-text content for assistant", () => {
-      const content: MessageContent = {
-        type: "attachment",
-        attachment: { type: "image", url: "..." },
-      } as any;
-
-      expect(() => toAiModelMessage(content, "assistant")).toThrow(
-        "Only text content is supported for role 'assistant'",
       );
     });
   });
